% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explainST.R
\name{explainST}
\alias{explainST}
\title{Explain a subgroup level taxa from Soil Taxonomy}
\usage{
explainST(x, format = c("text", "html"), viewer = TRUE)
}
\arguments{
\item{x}{a single subgroup, matching is exact and case insensitive}

\item{format}{output format: 'text' | 'html'}

\item{viewer}{show \code{format = 'html'} output in browser? default: \code{TRUE}}
}
\value{
a block of text, suitable for display in fixed-width font
}
\description{
Explain a subgroup level taxa from Soil Taxonomy
}
\note{
This function currently accepts only subgroup taxa. There are plans to extend to arbitrary levels of the hierarchy.
}
