% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSMR.R
\name{extractSMR}
\alias{extractSMR}
\title{Extract Soil Moisture Regime from Subgroup or Higher Level Taxon}
\usage{
extractSMR(taxon, as.is = FALSE, droplevels = FALSE, ordered = TRUE)
}
\arguments{
\item{taxon}{character. Vector of taxon names.}

\item{as.is}{Return character labels rather than an (ordered) factor? Default: \code{FALSE}}

\item{droplevels}{Drop unused levels? Default: \code{FALSE}}

\item{ordered}{Create an ordinal factor? Default: \code{TRUE}}
}
\value{
an (ordered) factor of Soil Moisture Regimes, or character vector when \code{as.is=TRUE}
}
\description{
Extract Soil Moisture Regime from Subgroup or Higher Level Taxon
}
\examples{
extractSMR(c("aquic haploxeralfs", "typic epiaqualfs", "humic inceptic eutroperox"))
}
