% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higherTaxaLevels.R
\name{level_to_taxon}
\alias{level_to_taxon}
\title{Get all taxa at specified level}
\usage{
level_to_taxon(level = c("order", "suborder", "greatgroup", "subgroup"))
}
\arguments{
\item{level}{character. One or more of "order", "suborder", "greatgroup", "subgroup"}
}
\value{
A character vector of taxa at the specified level
}
\description{
Convenience method for getting taxa from \code{ST_unique_list}
}
\examples{

# get all order and suborder level taxa

level_to_taxon(level = c("order","suborder"))

}
