% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newick.R
\name{newick_string}
\alias{newick_string}
\title{Generate Newick/New Hampshire Format Parenthetic Strings}
\usage{
newick_string(x = NULL, level = "suborder")
}
\arguments{
\item{x}{Optional: a taxon name to get children of.}

\item{level}{Level to build the tree at. One of \code{"suborder"}, \code{"greatgroup"}, \code{"subgroup"}. Defaults to \code{"suborder"} when \code{x} is not specified. When \code{x} is specified but \code{level} is not specified, \code{level} is calculated from \code{taxon_to_level(x)}.}
}
\value{
character. A single tree in parenthetical Newick or New Hampshire format.
}
\description{
This function generates Newick or New Hampshire format strings for a single tree. Taxa are assigned relative positions within their parent to indicate the order that they "key out."
}
\details{
The output from this function is a character string with parenthetical format encoding a single tree suitable for input into functions such as \code{ape::read.tree()}. Multiple trees can be combined together in the file or text string supplied to your tree-parsing function of choice.
}
\examples{
if (requireNamespace("ape")) {
  mytr <- ape::read.tree(text = newick_string(level = "suborder"))
  par(mar = c(0, 0, 0, 0))
  plot(mytr, "f")
}
}
