% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilTaxonomy-package.R
\docType{package}
\name{SoilTaxonomy-package}
\alias{SoilTaxonomy}
\alias{SoilTaxonomy-package}
\alias{SoilTaxonomy.env}
\title{SoilTaxonomy: A System of Soil Classification for Making and Interpreting Soil Surveys}
\description{
Taxonomic dictionaries, formative element lists, and functions related to the maintenance, development and application of U.S. Soil Taxonomy. Data and functionality are based on official U.S. Department of Agriculture sources including the latest edition of the Keys to Soil Taxonomy. Descriptions and metadata are obtained from the National Soil Information System or Soil Survey Geographic databases. Other sources are referenced in the data documentation. Provides tools for understanding and interacting with concepts in the U.S. Soil Taxonomic System. Most of the current utilities are for working with taxonomic concepts at the "higher" taxonomic levels: Order, Suborder, Great Group, and Subgroup.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ncss-tech/SoilTaxonomy}
  \item \url{https://ncss-tech.github.io/SoilTaxonomy/}
  \item \url{http://ncss-tech.github.io/SoilTaxonomy/}
  \item Report bugs at \url{https://github.com/ncss-tech/SoilTaxonomy/issues}
}

}
\author{
\strong{Maintainer}: Andrew Brown \email{andrew.g.brown@usda.gov}

Authors:
\itemize{
  \item Dylan Beaudette \email{dylan.beaudette@usda.gov}
}

}
\keyword{internal}
