% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilTaxonomyLevels.R
\name{SoilTaxonomyLevels}
\alias{SoilTaxonomyLevels}
\title{Get (Ordered) Factors based on Soil Taxonomy Key position}
\usage{
SoilTaxonomyLevels(
  level = c("order", "suborder", "greatgroup", "subgroup"),
  as.is = FALSE,
  ordered = TRUE
)
}
\arguments{
\item{level}{One of: \code{"order"}, \code{"suborder"}, \code{"greatgroup"}, \code{"subgroup"}}

\item{as.is}{Return just character labels rather than factor? Default: \code{FALSE}}

\item{ordered}{Create an ordinal factor? Default: \code{TRUE}}
}
\value{
an (ordered) factor or character vector (when \code{as.is=TRUE})
}
\description{
Get (Ordered) Factors based on Soil Taxonomy Key position
}
\examples{
SoilTaxonomyLevels("order")
SoilTaxonomyLevels("suborder")
}
