% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{ST}
\alias{ST}
\alias{ST_unique_list}
\title{Soil Taxonomy Hierarchy}
\format{
An object of class \code{data.frame} with 2665 rows and 9 columns.
}
\usage{
data(ST)
}
\description{
The first 4 levels of the US Soil Taxonomy hierarchy (soil order, suborder, greatgroup, subgroup), presented as a \code{data.frame} (denormalized) and a \code{list} of unique taxa.
}
\references{
Soil Survey Staff. 1999. Soil taxonomy: A basic system of soil classification for making and interpreting soil surveys. 2nd edition. Natural Resources Conservation Service. U.S. Department of Agriculture Handbook 436.
\url{https://www.nrcs.usda.gov/resources/guides-and-instructions/soil-taxonomy}

Soil Survey Staff. 2014. Keys to Soil Taxonomy, 12th ed. USDA-Natural Resources Conservation Service, Washington, DC.
\url{https://www.nrcs.usda.gov/resources/guides-and-instructions/keys-to-soil-taxonomy}
}
\keyword{datasets}
