% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{ST_family_classes}
\alias{ST_family_classes}
\title{Family-level Classes for Soil Taxonomy}
\format{
An object of class \code{data.frame} with 194 rows and 8 columns.
}
\usage{
data(ST_family_classes)
}
\description{
A database of family-level class names for Soil Taxonomy.
}
\references{
Soil Survey Staff. 2014. Keys to Soil Taxonomy, 12th ed. USDA-Natural Resources Conservation Service, Washington, DC.
\url{https://www.nrcs.usda.gov/resources/guides-and-instructions/keys-to-soil-taxonomy}
}
\keyword{datasets}
