% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{ST_features}
\alias{ST_features}
\title{Epipedons, Diagnostic Horizons, Characteristics and Features in Soil Taxonomy}
\format{
An object of class \code{data.frame} with 84 rows and 6 columns.
}
\usage{
data(ST_features)
}
\description{
A \code{data.frame} with columns "group", "name", "chapter", "page", "description", "criteria". Currently page numbers and contents are referenced to 12th Edition Keys to Soil Taxonomy and derived from products in the ncss-tech SoilKnowledgeBase repository (https://github.com/ncss-tech/SoilKnowledgeBase).
}
\references{
Soil Survey Staff. 2014. Keys to Soil Taxonomy, 12th ed. USDA-Natural Resources Conservation Service, Washington, DC.
\url{https://www.nrcs.usda.gov/wps/portal/nrcs/detail/soils/survey/class/taxonomy/?cid=nrcs142p2_053580}
}
\keyword{datasets}
