% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{ST_formative_elements}
\alias{ST_formative_elements}
\title{Formative Elements used by Soil Taxonomy}
\format{
An object of class \code{list} of length 4.
}
\usage{
data(ST_formative_elements)
}
\description{
A database of formative elements used by the first 4 levels of US Soil Taxonomy hierarchy (soil order, suborder, greatgroup, subgroup).
}
\references{
S. W. Buol and R. C. Graham and P. A. McDaniel and R. J. Southard. Soil Genesis and Classification, 5th edition. Iowa State Press, 2003.
}
\keyword{datasets}
