% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{ST_higher_taxa_codes_12th}
\alias{ST_higher_taxa_codes_12th}
\title{Letter Code Lookup Table for Position of Taxa within the Keys to Soil Taxonomy (12th Edition)}
\format{
An object of class \code{data.frame} with 3082 rows and 2 columns.
}
\usage{
data(ST_higher_taxa_codes_12th)
}
\description{
A lookup table mapping unique taxonomic Order, Suborder, Great Group and Subgroups to letter codes that denote their logical position within the Keys.
}
\details{
The lookup table has been corrected to reflect errata that were posted after the print publication of the 12th Edition Keys, as well as typos in the Spanish language edition.
}
\references{
Soil Survey Staff. 2014. Keys to Soil Taxonomy, 12th ed. USDA-Natural Resources Conservation Service, Washington, DC.
\url{https://www.nrcs.usda.gov/resources/guides-and-instructions/keys-to-soil-taxonomy}

Soil Survey Staff. 2014. Claves para la Taxonomía de Suelos, 12th ed. USDA-Natural Resources Conservation Service, Washington, DC.
\url{https://www.nrcs.usda.gov/Internet/FSE_DOCUMENTS/nrcs142p2_051546.pdf}
}
\keyword{datasets}
