% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{ST_higher_taxa_codes_13th}
\alias{ST_higher_taxa_codes_13th}
\title{Letter Code Lookup Table for Position of Taxa within the Keys to Soil Taxonomy (13th Edition)}
\format{
An object of class \code{data.frame} with 3153 rows and 2 columns.
}
\usage{
data(ST_higher_taxa_codes_13th)
}
\description{
A lookup table mapping unique taxonomic Order, Suborder, Great Group and Subgroups to letter codes that denote their logical position within the Keys.
}
\references{
Soil Survey Staff. 2022. Keys to Soil Taxonomy, 13th ed. USDA-Natural Resources Conservation Service.
\url{https://www.nrcs.usda.gov/resources/guides-and-instructions/keys-to-soil-taxonomy}
}
\keyword{datasets}
