% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{ST_SMR_13th}
\alias{ST_SMR_13th}
\title{Soil Moisture Regimes used in the Keys to Soil Taxonomy (13th Edition)}
\format{
An object of class \code{data.frame} with 3082 rows and 2 columns.
}
\usage{
data(ST_SMR_13th)
}
\description{
A data.frame containing taxon codes and the Soil Moisture Regime associated with that taxon.
}
\details{
Future revisions of this data set may additionally include the Soil Moisture Subclass.
}
\references{
S. W. Buol and R. C. Graham and P. A. McDaniel and R. J. Southard. Soil Genesis and Classification, 5th edition. Iowa State Press, 2003.
}
\keyword{datasets}
