% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higherTaxaCodes.R
\name{taxon_code_to_taxon}
\alias{taxon_code_to_taxon}
\title{Convert Taxon Code to Taxon}
\usage{
taxon_code_to_taxon(code)
}
\arguments{
\item{code}{A character vector of Taxon Codes}
}
\value{
A character vector of matching Taxon Names
}
\description{
Convert Taxon Code to Taxon
}
\examples{

taxon_code_to_taxon(c("ABC", "XYZ", "DAB", NA))

}
\seealso{
\code{\link{decompose_taxon_code}}, \code{\link{preceding_taxon_codes}}, \code{\link{taxon_to_taxon_code}}
}
