% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonTree.R
\name{taxonTree}
\alias{taxonTree}
\title{Create a \code{data.tree} Object from Taxon Names}
\usage{
taxonTree(taxon, root = "Soil Taxonomy", verbose = TRUE, ...)
}
\arguments{
\item{taxon}{A vector of taxon names}

\item{root}{Label for root node. Default: \code{"Soil Taxonomy"}; \code{NULL} for "unrooted" tree.}

\item{verbose}{Print tree output? Default: \code{TRUE}}

\item{...}{Additional arguments to \code{data.tree::as.Node.data.frame()}}
}
\value{
A data.tree object (invisibly). A text representation of the tree is printed when \code{verbose=TRUE}.
}
\description{
Create a \code{data.tree} Object from Taxon Names
}
\examples{
\dontshow{if (!inherits(requireNamespace("data.tree", quietly = TRUE), 'try-error')) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
taxonTree(c("hapludults", "hapludalfs"))
}
