% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isValidST.R
\name{isValidST}
\alias{isValidST}
\title{Check for valid order | suborder | greatgroup | subgroup.}
\usage{
isValidST(
  needle,
  level = c("tax_order", "tax_suborder", "tax_greatgroup", "tax_subgroup")
)
}
\arguments{
\item{needle}{vector of taxa}

\item{level}{single level of Soil Taxonomy hierarchy: tax_order, tax_suborder, tax_greatgroup, tax_subgroup}
}
\value{
logical vector, same lenght as needle
}
\description{
This function can be used to quickly test for valid levels within the
Soil Taxonomy hierarchy, a defined within the 13th edition of The Keys to Soil Taxonomy.
Matches are case insensitive.
}
\examples{

# note specfifcation of taxonomic level
isValidST('typic haploxeralfs', level = 'tax_subgroup')

}
