% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{ST_family_classes}
\alias{ST_family_classes}
\title{Family-level Classes for Soil Taxonomy}
\format{
An object of class \code{data.frame} with 188 rows and 7 columns.
}
\usage{
data(ST_family_classes)
}
\description{
A database of family-level class names for Soil Taxonomy.
}
\references{
Soil Survey Staff. 2014. Keys to Soil Taxonomy, 12th ed. USDA-Natural Resources Conservation Service, Washington, DC.
\url{https://www.nrcs.usda.gov/wps/portal/nrcs/detail/soils/survey/class/taxonomy/?cid=nrcs142p2_053580}
}
\keyword{datasets}
