% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higherTaxaCodes.R
\name{taxon_to_taxon_code}
\alias{taxon_to_taxon_code}
\title{Convert taxon name to taxon code}
\usage{
taxon_to_taxon_code(taxon)
}
\arguments{
\item{taxon}{A character vector of taxon names, case insensitive}
}
\value{
A character vector of matching taxon codes
}
\description{
Convert taxon name to taxon code
}
\examples{

taxon_to_taxon_code(c("Anhyturbels", "foo", "Cryaquands", NA))

}
\seealso{
\code{\link{decompose_taxon_code}}, \code{\link{preceding_taxon_codes}}, \code{\link{taxon_code_to_taxon}}
}
